rename: directory구조 변경
