+++ frontend/src/components/common/PcVersion/Asides/MemberCard.tsx
